build_targets: redesign target definition to be simpler
Currently, it is a bit hard for people without previous Python kowledge
to add new targets and platforms. The current design uses __init__ to
add any customizations, which is the correct way to do this in Python
code, but we are not really using this a Python code. This is suposed to
be a build target definition, which happens to use Python. Let's try to
simplify it.

It uses a metaclass to some magic handling of the methods defined in the
target class.

Signed-off-by: Filipe Laíns <lains@riseup.net>