build_targets: add include_files setting
Sometimes, like in the stm32f1-generic target, we want to include a
specific file in the preprocessor instead of a path. So this patch
introduces a new setting for that. include will add a path to the
include discovery path list, and include_files will add a file directly
to the propeocessor.

The stm32f1-generic taregt is currently broken because it passes the
config file via include, which will not do anything.

Signed-off-by: Filipe Laíns <lains@riseup.net>