targets: stm32f1-generic: fix passing the config file to the build system
This is currently broken as `include` will just append a path to the
include discovery path list, we want to include the config file instead.
This patch uses the newly introduced include_files setting to do that.
We are currently using platform_files to construct the file path, which
is obviously wrong, it should be target_files.

Signed-off-by: Filipe Laíns <lains@riseup.net>