configure: use static configs for the build system
Currently, all the targets and target families are defined in Python. I
designed a complex system to make it easier for people without much
knowledge of Python to add add new targets and dependencies. It is still
not great... and has a pretty high mantainence burden, as well a
frustration factor for newcomers if they run into issues. To hopefully
simplify things a little bit, I have changed this system to use static
TOML configs.
The new system is still not great, but I think is a slight improvement
over our previous one. There are still work to be done, and perhaps a
few internal redesigns as time progresses, but I think this is *fine*
for now...

Signed-off-by: Filipe Laíns <lains@riseup.net>